requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
generativeai==0.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-documentai==3.5.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.48.2
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.8
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.98.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
prettytable==3.16.0
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pyparsing==3.2.3
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.0.0
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.42
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.174
ultralytics-thop==2.0.15
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
